+++ config/test.exs
@@ -10,3 +10,9 @@ config :caylir, Caylir.TestHelpers.Graphs.InetsGraph,
   host: "localhost",
   pool: [ max_overflow: 0, size: 1 ],
   port: 9999
+
+
+config :caylir, Caylir.TestHelpers.Graphs.EnvGraph,
+  host: { :system, "CAYLIR_TEST_ENV_HOST" },
+  pool: [ max_overflow: 0, size: 1 ],
+  port: { :system, "CAYLIR_TEST_ENV_PORT" }