+++ test/helpers/graphs/env_graph.ex
@@ -1,0 +1,3 @@
+defmodule Caylir.TestHelpers.Graphs.EnvGraph do
+  use Caylir.Graph, otp_app: :caylir
+end